requirements.txt
bidict==0.23.1
blinker==1.9.0
click==8.1.8
colorama==0.4.6
dnspython==2.7.0
eventlet==0.39.1
Flask==3.1.0
flask-cors==5.0.1
Flask-SocketIO==5.5.1
Flask-SQLAlchemy==3.1.1
greenlet==3.1.1
h11==0.14.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
pydub==0.25.1
python-decouple==3.8
python-dotenv==1.1.0
python-engineio==4.11.2
python-socketio==5.12.1
simple-websocket==1.1.0
SpeechRecognition==3.14.2
SQLAlchemy==2.0.40
typing_extensions==4.13.0
Werkzeug==3.1.3
wsproto==1.2.0
